+++ src/main/java/com/badbones69/crazycrates/tasks/crates/CrateManager.java
@@ -406,11 +406,11 @@ public class CrateManager {
 
                 final PluginManager server = this.plugin.getServer().getPluginManager();
 
-                final boolean isOldSystemEnabled = this.config.getProperty(ConfigKeys.use_old_permission_system);
+                final boolean isNewSystemEnabled = this.config.getProperty(ConfigKeys.use_new_permission_system);
 
-                final String node = isOldSystemEnabled ? "crazycrates.open." + crateName : "crazycrates.deny.open." + crateName;
-                final String description = isOldSystemEnabled ? "Lets you open " + crateName : "Prevents you from opening " + crateName;
-                final PermissionDefault permissionDefault = isOldSystemEnabled ? PermissionDefault.TRUE : PermissionDefault.FALSE;
+                final String node = isNewSystemEnabled ? "crazycrates.deny.open." + crateName : "crazycrates.open." + crateName;
+                final String description = isNewSystemEnabled ? "Prevents you from opening " + crateName : "Lets you open " + crateName;
+                final PermissionDefault permissionDefault = isNewSystemEnabled ? PermissionDefault.FALSE : PermissionDefault.TRUE;
 
                 if (server.getPermission(node) == null) {
                     final Permission permission = new Permission(node, description, permissionDefault);