+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Added:
 - Added a toggle, which allows you to revert the chance made previously.
   - crazycrates.open.<crate_name> -> crazycrates.deny.open.<crate_name>
-- The configuration option can be found in the `config.yml` i.e. `root.use-old-permission-system`
+- The configuration option can be found in the `config.yml` i.e. `root.use-new-permission-system`
   - This option likely won't be here forever, one of the systems will be default. You'll be notified when that happens.